build: Don't use block mode by default, allow it to be chosen
Currently this uses the same logic as backuptool (but inverted)

i.e. block mode will be disabled in all cases where backuptool is enabled
which is the current default for all Omni builds.

Block mode is not needed now, and may never be needed for custom projects,
as block mode is preparation for dm-verity, which is mostly pointless on
devices with unlocked bootloaders since the kernel already has broken
the chain of trust.

TODO:  Determine if Nexus 9 *requires* blockmode.  If it does, this
may need to be reworked to allow the device to specify blockmode.

Note: set TARGET_USES_BLOCK_MODE := true to use block mode

Change-Id: I36725b44f4066a43ec7e1615bd4b072f7eb3db28