armv7-a-neon: Generate proper instructions for cortex-a53
This is needed for devices with armv8 SoCs, which are stuck
on 32bit systems for various reasons (bootloader, hypervisor, etc).

An example for this is the Galaxy Note 4 Exynos variant with the
Exynos 5433 Octa Soc. Samsung did not ship any 64bit builds for this
device, thus we're unable to use a proper armv8 configuration, because
Android only supports this on an arm64 target.

Change-Id: Ib1018654d215494353d5af59751249ee6dbcf40e